+++ TablePro/Views/ConnectionForm/Panes/SSLPaneView.swift
@@ -14,7 +14,8 @@ struct SSLPaneView: View {
             sslMode: $coordinator.ssl.mode,
             sslCaCertPath: $coordinator.ssl.caCertPath,
             sslClientCertPath: $coordinator.ssl.clientCertPath,
-            sslClientKeyPath: $coordinator.ssl.clientKeyPath
+            sslClientKeyPath: $coordinator.ssl.clientKeyPath,
+            sslClientKeyPassphrase: $coordinator.ssl.clientKeyPassphrase
         )
     }
 }